Fixed generation of multi-line source fields in a cell
Using equal? to check for last element within array now.
This also works, if the array elements are not unique,
which commonly is the case, as an array of code lines could
contain multiple empty lines. The test for the last element
did on empty lines, if the last line was also empty.